requirements.txt
azure-mgmt-resource
azure-mgmt-datalake-store
azure-storage-file-datalake
azure-identity
python-dotenv